Require node >= 0.8.19.
This project is using peerDependencies to ensure that the Handlebars
dependency is setup properly (and can be overridden by our consumers).
In practice this works wonderfully, but a few users with older Node.js
and/or NPM versions simple get a broken build. This is because Node.js <
0.8.19 did not have the ability to handle peerDependencies specified in
the `package.json`.

This changes to ensure that at least 0.8.19 (well over a year old) is used.